Add "self destruct" to the Science Drone
If the player repairs, then attacks & boards the Beagle with a drone carrier, they will carry it. Since it cannot be made to pass the current flight check warnings (after landing on a planet), a self-destruct attribute prevents the player from capturing it and ending up on a planet where they cannot sell it, or take off without being forced to first park or disown it.