Adding the TAS
Tactical Assault System. (aka a pulley, winch, and cable system to allow your crew fast access to all parts of your ship. This is intended to allow unpredictable and unusual counter-boarding tactics.

(for an illustration, read Ender's Game and the unusual tactics he could use when they gave him ropes to use in the zero-grav arena and how much of an advantage they were.)