Lowered the weapon capabilities of the Petrel
The Petrel had far too much weapon space compared to other fighters. With this reduction, the Petrel maintains its positions as having the most weapon space (as well as the most weapon hardpoints), but there isn't as absurd a gap. If new fighter categories existed, then the Petrel's stats could be reverted and it could be moved into the new category.