feat: Removed some Sestor ships from the Alnilam fleets to make bombing Zenith easier for those without a cloak
Largely focused on reducing the number of Detainers that the Sestor ships had. Now only the Tek Far 71 - Lek has a Detainer, whereas before 7 separate ships had Detainers, making it nearly impossible to land on Zenith. Now, it's possible to either get lucky with the Tek Far 71 being placed far from the planet, or you can thrust to the system edge to avoid the Detainer fire then pick some of the Sestor ships off and come back in to land.