Renaming the Gascraft to the Puffin
Replaced all instances of "gascraft" with "puffin". Instances are located in: remnant missions.txt, remnant ships.txt, remnant.txt, as well as the sprite & thumbnail.

Also relocated a copy of the gascraft description to the "deprecated outfits.txt" file and added a compatibility mission to give anyone with the gascraft event the appropriate puffin event.

This also relocates the puffin ship info to be in proper alphabetical order according to its new name.