feat: Switch off fast-forward if the player is not in flight (#5148)
This is so that the behavior of fast-forward is consistent between panels (#5038). This also gives the player extra time to react when coming out of a boarding panel or on enter conversation by turning fast-forward off if it was on when entering the panel.

The map panel and player info panel are exceptions to this, as they are panels that are commonly accessed and can be accessed at any time while in flight.