feat(content): Removed portrait images from spaceport news (#5187)
Spaceport portraits of people were an experimental idea that has received generally negative feedback. See #4325 or the comments on #3389 after the news system was added.
This PR removes the use of the portraits, but keeps the portrait mechanic in game, as we may find a use for it later that doesn't involve images of people, or plugins may still want to make use of the portraits of people.