feat(mechanics): Overhaul jamming and tracking to make them more fun and realistic (#4204)
This PR aims to make missile tracking and jamming much more strongly affected by the characteristics of the target ship and the missile's tracking and homing methods, rather than being something that almost always works no matter what. This is more fun, rewards the player for equipping their ship appropriately, and makes missiles a bit more likely to miss their targets on average.

To that effect, it implements the following technical changes to jamming and tracking:

- Double the "resolution" of missile tracking by checking for lock twice as frequently
- When a missile has lost its lock, zero out its base chance to regain it
- Attenuate the effects of radar jamming proportional to the distance between the jammer and the
   missile; more jamming means that radar-guided missiles can become jammed at longer ranges
- Make jammed radar-guided missiles have a chance to "go haywire", proportional again to the
   strength of the jamming compared to the strength of the missile's radar tracking modified by the
   distance to the jammer
- Increase the reliability of infrared tracking when close to the target, mimicking the real-world
   phenomenon stemming from the wavelengths of infrared light being easier to distinguish with
   greater distance
- Slightly reduce the ship weight at which optical tracking becomes maximally effective to
   compensate for tracking being a bit less accurate now

Taken together, these changes mean that running a cool ship actually causes IR missiles to miss a lot of the time, and installing even a Small Radar Jammer is enough to make a fraction of Sidewinders miss against a maneuvering target. Torpedoes miss a little bit more against large ships, and are now almost useless against small maneuvering ships.

This PR also makes the combination of radar and IR tracking in a single missile make sense: radar tracking is used for long-range effectiveness, and the missile switches to IR when it's close to the target. Some of the best real-world missiles do this, too.

closes #1865
closes #3912