fix(content): Adjust human guided missiles to adjust to jamming changes (#4212)
With the jamming and tracking changes in #4204, I've slightly adjusted the Torpedoes, Sidewinder, and Meteor.

The humble Meteor finally has a useful battle role: it is now a short-range bombardment weapon that is optimally fired from fighters against medium-to-large-sized ships, but can be used against small maneuvering ships when fired by a skillful pilot (i.e. the player).

However, it's still just as forgiving to new players--in fact maybe even more so: By worsening its tracking, it is now easier to dodge in a small ship. Its reduced range means that it won't come from out of nowhere. Reduced homing mean that when it misses, it never comes back around for another attack run.

The sidewinder's range was nearly cut in half. It's now longer-ranged than the meteor, but shorter than the Torpedo. It's essentially just an anti-fighter weapon now, and very effective in this role. However it can be pressed into service against warships in a pinch (it is very expensive to do this, just like when the U.S. Navy uses a 4 million dollar Standard Missile against ships instead of than fighters or other missiles, which are its intended targets).

The Torpedoes are now the go-to long range anti-capital ship weapons. The changes to tracking make them virtually worthless against small maneuvering ships, but they are devastating to capital ships.