fix(mechanics): Increased radar jamming and reduced tracking properly increase haywire chance
The weaker the radar tracking on a projectile, the more likely it is to go haywire.
The higher the jamming on the projectile's target, the more likely it is to go haywire.
The closer the projectile is to the target with jamming, the more likely it is to go haywire.
And vice versa.

The old equation was a little overly complex and contained some errors that not only resulted in strange behavior when changing certain values value (e.g. a tracking of just above 0 and just below 1 had the same results if you had a jamming of 2), but I believe also did the opposite of what was intended.

Also some housekeeping.