docs(readme): Identify community-run Discord server as a source of community engagement (#5743)
* Replacing Main Page link to google groups with discord link

We really shouldn't be directing people to the google groups from the front page of the Endless Sky repository. None of the core developers or content creators frequent there, whereas they virtually all are on the discord. Let's direct people interested in contributing to where they can actually connect with us.

* MCO suggested wording change

"discussion discord" to "community discord"

Co-authored-by: M*C*O <mcofficer@gmx.de>

* Update README.md

Added link to the steam discussion pages.

Added an "Archival" section with a link to the google groups, as well as an alternative link.

* Update README.md

missing closing paranthesis

* Clarifying the statement

So far (19h into the poll), no one has been able to access the google groups while:
a) logged into a google account that has never visited it before. (26 people have tried) or
b) not logged in. (22 people have tried)

Of the 15 people who have previously accessed it, only 12 could access it during this poll.

* docs(readme): update wording

Co-authored-by: M*C*O <mcofficer@gmx.de>
Co-authored-by: tehhowch <tehhowch@users.noreply.github.com>