feat(uuids): Add a container comparator to support unique-by-UUID containers
There isn't an immediately obvious use for UUID-keyed collections, but future work could benefit from it, such as making the AI::orders collection persistent across save reloads.
In general, only collections that need to be serialized to disk would benefit, as the pointer to a UUID-enabled object will be stable throughout the session.