feat(mechanics): Add basic support for ships whose sprites contain holes / disjoint regions (#3621)
Ship AI does not account for the possibility that aiming at the "center" of a body may not actually hit said object. Content creators are responsible for not abusing mechanics (and are of course free to do so in content they do not wish to upstream).

* Mask traces multiple outlines.
Allows holes and independent shapes.

* Improve shape of simplified outline.
Uses squared area instead of squared line distance to decide when to
stop simplifying, resulting in an outline closer to the original image.
An area that produces a similar number of points was chosen.

* feat(masks): add some error logging

* refactor(masks): pass around result vectors by reference
 - nullptr isn't an option, so use reference semantics
 - Update a few comments to be more accurate

* refactor(masks): Use RVO in Simplify & mark unused outline capacity as free

* chore: cleanup
Still need to determine whether the `isAlone` check makes sense -- it should
perhaps be a `break`, or ignored if there are well-formed outlines already
stored, etc. As is, the second `if(isAlone) break;` statement will not be
executed, but is kept for posterity pending resolution to the first statement.

* fix(leaks): don't segfault if a maskless ship is somehow exploding
It couldn't take damage, but it could e.g. self-destruct due to firing hull.

Co-authored-by: tehhowch <tehhowch@users.noreply.github.com>