fix(content): Ensure all sprites have even dimensions to avoid induced blurriness (#6248)
This PR fixes issue #5456 and removes all instances of ships, thumbnails and outfit images having odd-numbered image dimensions, thereby making them show up blurry in-game. Hardpoints and effects were chosen to not be modified in this PR, as they are small enough/only show up for brief enough that there wouldn't be a noticeable difference in quality.