fix(content): Define and sort illegal cargoes better (#6243)
Overhauled and sharpened the definitions of "illegal" and "highly
illegal" for commodities. Now the highly illegal items are the kinds of
things that are actually considered really really bad, like hard drugs,
date rape pills, assassination equipment, tools for evading government
scrutiny, biological agents, weapons of mass destruction and their
precursors, and the like. This should sharpen the distinction between
the "just illegal" jobs and the "really illegal" jobs. This is important
because it's possible to feel like an edgy Han Solo-like rebel by
smuggling untaxed liquor and counterfeit artwork, but if you start
smuggling roofies and bioweapons, you're pretty solidly in "I'm a bad
guy" now territory".