feat(content): add minor Deep flavor missions (#6234)
This adds a series of missions that begin in Helheim, move to Norn, and eventually take the player to Alfheim. Doing the entire series of quests nets the player 500,000 credits. However there is a 30 day and 60 day wait between some missions, meaning that it would take at least 90 in-game days to complete.

To start the mission, head to Helheim spaceport.

Co-authored-by: Terin <4471575+Terin@users.noreply.github.com>
Co-authored-by: Wizrad <onlyw1zrad@gmail.com>
Co-authored-by: MasterOfGrey <60949828+MasterOfGrey@users.noreply.github.com>
Co-authored-by: Nate Graham <pointedstick@zoho.com>