refactor(internal): Generate list of equipped weapons on demand (#6504)
Using a class field to track the installed weapon outfits while parsing the ship template into a class object is easier to do incorrectly, adds complexity, and also increases the `Ship` class size unnecessarily for most of the game's purposes. Instead, use the `Armament` class and generate this list when needed based on what was parsed already.

* Generate the list of equipped weapon-outfits on-demand

* Prevent showing 0 outfits for equipped, but not installed, weapons.

* Add warning for installed weapons with no slot.

* PR feedback

* std::move the warning-string instead of copying it.

Also removed 2 characters that were duplicates.

* Use the modelName and name getters instead of the fields.

This is more future-proof, in case we decide to store the modelNames and
name differently in the future, or in case we decide to change how those
names are displayed in the future.

* Use VariantName instead of ModelName.

* Remove non-needed std::moves and some code improvement