docs: Update PR template to point to the assets repo
Previously we simply asked that a link be provided to the assets, but this has led to assets being scattered across various PRs as no effort has been made to recentralize the assets into a single location. Now that we have the assets repo (more) up to date, make it a requirement that a PR be made to the assets repo instead of just providing a link to some random location where the assets are held.