feat(AI): Improve coward and appeasing ship behavior (#6421)
Ships with these personalities previously triggered at 35% health remaining until disabled. This has been changed to 50%, as the original 35% was too close to the standard retreat health threshold of 25% to have any real effect on behavior. Coward should also now break all targets and not look for a new one when fleeing.